+++ __tests__/__system__/__scripts__/imperative_zosmf_create_profile.sh
@@ -1,0 +1,9 @@
+#!/bin/bash
+set -e
+
+echo "================ZOSMF HELP==============="
+zowe profiles create zosmf $*
+if [ $? -gt 0 ]
+then
+    exit $?
+fi