+++ packages/zosmf/__tests__/ZosmfSession.unit.test.ts
@@ -10,7 +10,7 @@
 */
 
 import { ZosmfSession } from "../src/api/ZosmfSession";
-import { Session } from "@zowe/imperative";
+import { Session, ImperativeExpect } from "@zowe/imperative";
 
 describe("zosmf utils", () => {
     it("should create a session object", () => {
@@ -23,5 +23,34 @@ describe("zosmf utils", () => {
             rejectUnauthorized: "fake"
         });
         expect(session.ISession).toMatchSnapshot();
+    });
+    it("should fail to create a session object when username, and password are not present", () => {
+        let error;
+        try {
+            const session: Session = ZosmfSession.createBasicZosmfSession({
+                host: "fake",
+                port: "fake",
+                rejectUnauthorized: "fake"
+            });
+        } catch (err) {
+            error = err;
+        }
+        expect(error.toString()).toContain("Must have user & password OR base64 encoded credentials");
+    });
+
+    it("should fail to create a session object when host is not present", () => {
+        let error;
+        try {
+            const session: Session = ZosmfSession.createBasicZosmfSession({
+            port: "fake",
+            user: "fake",
+            password: "fake",
+            auth: "fake",
+            rejectUnauthorized: "fake"
+            });
+        } catch (err) {
+            error = err;
+        }
+        expect(error.toString()).toContain("Required parameter 'hostname' must be defined");
     });
 });