+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 
 All notable changes to the Zowe CLI package will be documented in this file.
 
+## `6.12.0`
+
+- Make username, password, and host optional on z/OSMF profiles and update profile creation doc to reflect the change.
+- Don't overwrite files when downloading spool output from job with duplicate step names.
+
+## `6.11.0`
+
+- Add preserve-original-letter-case option for download to keep generated folders and files in original uppercase.
+
+## `6.10.3`
+
+- Update Migrate and Recall data set APIs to have a base handler function.
+
+## `6.10.2`
+
+- Update Imperative to 4.6.
+- Update top-level doc links in help description.
+
 ## `6.10.1`
 
 - Update Imperative dependency to fix vulnerability.